requirements.txt
flask
gunicorn
numpy
tensorflow-cpu==2.10.0